forex-dashboard/service/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
APScheduler==3.10.4
python-dotenv>=1.1.0
